Add lograge gem for log formatting
Add lograge to format our logs for shipping off to cloudwatch. This includes
adding logstash-event gem which is the formatter `Lograge::Formatters::Logstash`
that we are using to create json logs.

Also adds configuration to config/environments/production.rb so that we only
format the logs for our production environment.